Partially roll back the last RC - a compromise
With this release, you extend Phrases, not strings. However,
the Phrases are checked for an "id" property, which is used if
the actual comparison for equality doesn't work out. The advantage of
this is that you can have extension of phrases from the same package
without string IDs, but you can include them to remove the browserification
issues outside the package.